FEAT (Roadmap) : splits comp into Desktop and Mobile views
